Add files via upload
In this project, we have implemented various machine-learning algorithms on different types of datasets. The project includes supervised learning (classification) on the Titanic dataset, supervised learning (regression) on the medical price dataset, and unsupervised learning using the K-Means Clustering and Principal Component Analysis algorithms on the Iris dataset. The objective of this project was to explore and apply different machine-learning techniques to gain insights from the datasets and evaluate the performance of the algorithms